modules/rosidl: Consolidate message file generations
This commit implements a new version of gen_file().  This version
takes a list of .msg and .srv files, return generated header and
source files.  All conversion is done in the functions and callers
doesn't have to care about it.  All it needs is a list of message
files it wants to use.

The callers must depend on the generated header files, and build
generated source files in addition to its own files.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>