modules/microxrcedds-client: Convert cmake configuration to Kconfig
Move all configuration parameter in CMakeLists.txt to Kconfig.  The
original config.h was generated from config.h.in in the source. But
With this commit we host uxr/client/config.h in this repository and
let the module reference it.

Currently we do not support any transport except custom one.  We need
to upstream transports for Zephyr in order to do so.  The menus for
other unsupported transports are there to show off how it supposed to
be.

Some configurations are not included, namely multi-thread, shared
memory, qos matching.

PROJECT_VERSION* are removed since it seems none is using them.

Hard liveliness check is changed to default `ON` instead of
`OFF`. This is because micro-ROS Agent dies without it, as described
in the previous commit.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>